getDataSensores/requirements.txt
flask
requests 
numpy 
pandas 
python-dotenv
paho-mqtt
pika
flask
flask-sqlalchemy
eventlet
Flask-MQTT
Flask-SocketIO
mysqlclient
pyjwt